fix: moving reservation action logic to Reservations module
